Change Methods

+++ GPA_Calculator/Repositories/SubjectRepository.cs
@@ -14,7 +14,7 @@ namespace GPA_Calculator.Repositories
         Task<List<SubjectEntity>> GetTop3EasySubject();
 
         Task<List<SubjectEntity>> GetLast3HardSubject();
-        Task<double> GetGPA(double studentId);
+        Task<double> GetGPA(int studentId);
 
     }
 
@@ -108,7 +108,7 @@ namespace GPA_Calculator.Repositories
         }
 
 
-        public async Task<double> GetGPA(double studentId)
+        public async Task<double> GetGPA(int studentId)
         {
             var gpa = 0.0;
             var GP = 0.0;
